Complete lesson 23.

+++ app/Http/Controllers/JobController.php
@@ -3,7 +3,11 @@
 namespace App\Http\Controllers;
 
 use App\Models\Job;
+use App\Models\User;
+use App\Providers\AppServiceProvider;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Auth;
+use Illuminate\Support\Facades\Gate;
 
 class JobController extends Controller
 {
@@ -59,7 +63,7 @@ class JobController extends Controller
     }
 
     public function destroy(Job $job) {
-        // authorization... on hold
+        Gate::authorize('edit-job', $job);
 
         $job->delete();
